% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2triples.R
\name{dataframe2triples}
\alias{dataframe2triples}
\title{converts a rectangular data frame to a list of triple definitions}
\usage{
dataframe2triples(data, root = "_:data")
}
\arguments{
\item{data}{the data frame to be translated}

\item{root}{the node, potentially a blank node, to which all observations are linked}
}
\value{
a list of character vectors
}
\description{
Returns a list of triples definitions suitable for \link[rsRDF]{add.edges}
from a single input data frame. Is linked back to a node given by the
parameter \code{root}.
}
\examples{
A=data.frame(x=seq(1,5,1))
A$y1=sin(A$x)
A$y2=cos(A$x)

dataframe2triples(A) |> lapply(paste,collapse=" ") |> unlist()
}
