% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeEncoder.R
\name{dataframeEncoder}
\alias{dataframeEncoder}
\title{converts a rectangular data frame to a list of triple definitions}
\usage{
dataframeEncoder(data, root = "_:data", tag = NULL, read = NULL)
}
\arguments{
\item{data}{the data frame to be translated}

\item{root}{the node, potentially a blank node, to which all observations are linked}

\item{read}{}
}
\value{
a list of character vectors
}
\description{
Returns a list of triples definitions suitable for \link[rsRDF]{add.edges}
from a single input data frame. Is linked back to a node given by the
parameter \code{root}.
}
\examples{
A=data.frame(x=seq(1,5,1))
A$y1=sin(A$x)
A$y2=cos(A$x)

dataframeEncoder(A) |> lapply(paste,collapse=" ") |> unlist()
}
