% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zulu.R
\name{Zulu}
\alias{Zulu}
\title{returns the current time with milliseconds}
\usage{
Zulu(spaces = T, short = F, time.only = F)
}
\arguments{
\item{spaces}{false, all spaces are replaced by underscores}

\item{short}{if true, simply reports the hours, minutes and seconds in UTC
preceded by the day of the week}

\item{time.only}{if true, the date is suppressed and just the time printed}
}
\value{
a character string with the current time in UTC
}
\description{
Returns the current time in millisections for the UTC (GMT) time zone. Note
that the returned value is a character format for which order relations are
defined (including \code{max()} and \code{min()}) but not further arithmetic functions
(e.g. \code{mean()}).
}
\details{
A final Z for \code{Zulu} is added to stop RDF coercing the output to a (truncated)
date format. Use \code{short=T} or \code{time.only=T} for a short form.
}
\examples{
a1=Zulu()
a1

a2=Zulu()
a2

a1 < a2
a1 > a2

class(a1)

Zulu(F)
Zulu(short=T)
Zulu(time,only=T)
}
