% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach.R
\name{Reach}
\alias{Reach}
\title{connects to an external community hosting RDF store}
\value{
a data frame or NULL
}
\description{
connects to an external community hosting RDF store

connects to an external community hosting RDF store
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{w}}{the \link[rsAdmin]{Remote} object supporting the datastore connection}

\item{\code{endpoint}}{the URL of the remote triple store}

\item{\code{userid}}{the userid of the current client}

\item{\code{created}}{the time of creation of this R6 object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reach-new}{\code{Reach$new()}}
\item \href{#method-Reach-print}{\code{Reach$print()}}
\item \href{#method-Reach-myName}{\code{Reach$myName()}}
\item \href{#method-Reach-clone}{\code{Reach$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reach-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reach-new}{}}}
\subsection{Method \code{new()}}{
initialises a connection object to a remote server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reach$new(
  endpoint = "http://localhost:8080/rdf4j-server/repositories/01",
  userid = "007"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{the URL of the remote RDF triple store}

\item{\code{userid}}{the clients userid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reach-print"></a>}}
\if{latex}{\out{\hypertarget{method-Reach-print}{}}}
\subsection{Method \code{print()}}{
prints information about the connection and userid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reach$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reach-myName"></a>}}
\if{latex}{\out{\hypertarget{method-Reach-myName}{}}}
\subsection{Method \code{myName()}}{
checks the current given user name against the remote database. If found
the remote entry is printed. If not a value of NULL is return
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reach$myName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reach-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reach-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reach$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
