% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsComms.R
\name{rsComms}
\alias{rsComms}
\title{connects to an external community hosting RDF store}
\value{
a data frame or NULL

a data frame or NULL
}
\description{
connects to an external community hosting RDF store

connects to an external community hosting RDF store
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{w}}{the \link[rsAdmin]{Remote} object supporting the datastore connection}

\item{\code{endpoint}}{the URL of the remote triple store}

\item{\code{userid}}{the userid of the current client}

\item{\code{created}}{the time of creation of this R6 object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rsComms-new}{\code{rsComms$new()}}
\item \href{#method-rsComms-print}{\code{rsComms$print()}}
\item \href{#method-rsComms-myName}{\code{rsComms$myName()}}
\item \href{#method-rsComms-collaboration}{\code{rsComms$collaboration()}}
\item \href{#method-rsComms-report}{\code{rsComms$report()}}
\item \href{#method-rsComms-clone}{\code{rsComms$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-new"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-new}{}}}
\subsection{Method \code{new()}}{
initialises a connection object to a remote server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$new(
  endpoint = "http://localhost:8080/rdf4j-server/repositories/01",
  userid = "007"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{the URL of the remote RDF triple store}

\item{\code{userid}}{the clients userid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-print"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-print}{}}}
\subsection{Method \code{print()}}{
prints information about the connection and userid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-myName"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-myName}{}}}
\subsection{Method \code{myName()}}{
checks the current given user name against the remote database. If found
the remote entry is printed. If not a value of NULL is return
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$myName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-collaboration"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-collaboration}{}}}
\subsection{Method \code{collaboration()}}{
Lists or requests memberships of a collaboration. Where operations are not
logically possible, an explanatory message is printed to the screen.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$collaboration(action = "list", target = "01")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{the desired action: \code{list} prints a list of all collaborations
and whether the current user is already a member, whilst \code{join} requests to
join a collaboration with the id \code{target} below.}

\item{\code{target}}{the id of a collaboration to join should action be set to \code{join}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-report"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-report}{}}}
\subsection{Method \code{report()}}{
lists or creates reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$report(
  action = "list",
  id = "01",
  info = "this is a report",
  triples = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{one of \code{create} to create a new report or \code{list} to list
all available reports, or \code{post-data} or \code{read-data} to post or read
data.}

\item{\code{id}}{the report ID if creating a report}

\item{\code{info}}{an info string if creating a report}

\item{\code{triples}}{triples to be posted for an \code{add-data} action}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsComms-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rsComms-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rsComms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
